lesson17/requirements.txt
pytest==7.0.1
allure-pytest==2.9.45
pytest-rerunfailures==10.2